requirements.txt
pymongo
python-dotenv
openai
scikit-learn
numpy
certifi
